Moving RFBQRequest to public file
